[stepper] Use next.js as per #7759 (#4799)
