+++ docs/src/pages/demos/stepper/HorizontalNonLinearStepper.js
@@ -2,78 +2,43 @@
 
 import React, { Component } from "react";
 import PropTypes from "prop-types";
-import { withStyles, createStyleSheet } from "material-ui/styles";
+import { withStyles } from "material-ui/styles";
 import { Step, Stepper, StepLabel, StepButton } from "material-ui/Stepper";
 import Button from "material-ui/Button";
 import Typography from "material-ui/Typography";
 
-const styleSheet = createStyleSheet("HorizontalLabelPositionBelowStepper", theme => ({
+const styles = theme => ({
   root: {
     width: '90%'
   },
   button: {
     marginRight: theme.spacing.unit,
   },
-  backButton: {
-    marginRight: theme.spacing.unit
-  },
   completed: {
     display: 'inline-block',
   },
-}));
+});
 
-class HorizontalNonLinearAlternativeLabelStepper extends Component {
+class HorizontalNonLinearStepper extends Component {
   state = {
     activeStep: 0,
-    completed: new Set(),
-    skipped: new Set(),
+    completed: {},
   };
+
+  completedSteps() {
+    return Object.keys(this.state.completed).length;
+  }
 
   totalSteps() {
     return this.getSteps().length;
-  }
-
-  isStepComplete(step) {
-    return this.state.completed.has(step);
-  }
-
-  completedSteps() {
-    return this.state.completed.size;
-  }
-
-  allStepsCompleted() {
-    return this.completedSteps() === this.totalSteps() - this.skippedSteps();
   }
 
   isLastStep() {
     return this.state.activeStep === this.totalSteps() - 1;
   }
 
-  isStepOptional(step) {
-    return step === 1;
-  }
-
-  isStepSkipped(step) {
-    return this.state.skipped.has(step);
-  }
-
-  handleSkip = () => {
-    const activeStep = this.state.activeStep;
-    if (!this.isStepOptional(activeStep)) {
-      // You probably want to guard against something like this - it should never occur unless someone's actively
-      // trying to break something.
-      throw new Error("You can't skip a step that isn't optional.")
-    }
-    const skipped = new Set(this.state.skipped.values());
-    skipped.add(activeStep);
-    this.setState({
-      activeStep: this.state.activeStep + 1,
-      skipped,
-    });
-  }
-
-  skippedSteps() {
-    return this.state.skipped.size;
+  allStepsCompleted() {
+    return this.completedSteps() === this.totalSteps();
   }
 
   handleNext = () => {
@@ -82,7 +47,7 @@ class HorizontalNonLinearAlternativeLabelStepper extends Component {
     if (this.isLastStep() && !this.allStepsCompleted()) {
       // It's the last step, but not all steps have been completed - find the first step that has been completed
       const steps = this.getSteps();
-      activeStep = steps.findIndex((step, i) => !this.state.completed.has(i));
+      activeStep = steps.findIndex((step, i) => !(i in this.state.completed));
     } else {
       activeStep = this.state.activeStep + 1;
     }
@@ -93,7 +58,7 @@ class HorizontalNonLinearAlternativeLabelStepper extends Component {
 
   handleBack = () => {
     this.setState({
-      activeStep: this.state.activeStep - 1
+      activeStep: this.state.activeStep - 1,
     });
   };
 
@@ -104,21 +69,18 @@ class HorizontalNonLinearAlternativeLabelStepper extends Component {
   }
 
   handleComplete = () => {
-    const completed = new Set(this.state.completed.values());
-    completed.add(this.state.activeStep);
+    const completed = this.state.completed;
+    completed[this.state.activeStep] = true;
     this.setState({
       completed,
     });
-    if (!this.allStepsCompleted()) {
-      this.handleNext();
-    }
+    this.handleNext();
   }
 
   handleReset = () => {
     this.setState({
       activeStep: 0,
-      completed: new Set(),
-      skipped: new Set(),
+      completed: {},
     });
   };
 
@@ -144,23 +106,14 @@ class HorizontalNonLinearAlternativeLabelStepper extends Component {
 
     return (
       <div className={classes.root}>
-        <Stepper alternativeLabel nonLinear activeStep={activeStep}>
-          {steps.map((label, step) => {
-            const props = {};
-            if (this.isStepOptional(step)) {
-              props.optional = true;
-            }
-            if (this.isStepSkipped(step)) {
-              props.completed = false;
-            }
-            return (
-              <Step key={step} {...props}>
-                <StepButton onClick={this.handleStep(step)} completed={this.isStepComplete(step)}>
-                  {label}
-                </StepButton>
-              </Step>
-            );
-          })}
+        <Stepper nonLinear activeStep={activeStep}>
+          {steps.map((label, i) => (
+            <Step key={i}>
+              <StepButton onClick={this.handleStep(i)} completed={this.state.completed[i]}>
+                {label}
+              </StepButton>
+            </Step>
+          ))}
         </Stepper>
         <div>
           {this.allStepsCompleted()
@@ -179,13 +132,8 @@ class HorizontalNonLinearAlternativeLabelStepper extends Component {
                   < Button raised color="primary" onClick={this.handleNext} className={classes.button}>
                     Next
                   </Button>
-                  {this.isStepOptional(activeStep) &&
-                    <Button raised color="primary" onClick={this.handleSkip} className={classes.button}>
-                      Skip
-                    </Button>
-                  }
                   {activeStep !== steps.length && (
-                    this.state.completed.has(this.state.activeStep)
+                    this.state.completed[this.state.activeStep]
                       ? <Typography type="caption" className={classes.completed}>Step {activeStep + 1} already completed</Typography>
                       : <Button raised color="primary" onClick={this.handleComplete}>
                         {this.completedSteps() === this.totalSteps() - 1
@@ -203,8 +151,8 @@ class HorizontalNonLinearAlternativeLabelStepper extends Component {
   }
 }
 
-HorizontalNonLinearAlternativeLabelStepper.propTypes = {
+HorizontalNonLinearStepper.propTypes = {
   classes: PropTypes.object
 };
 
-export default withStyles(styleSheet)(HorizontalNonLinearAlternativeLabelStepper);
+export default withStyles(styles)(HorizontalNonLinearStepper);
